NS running in series on multiple fits

+++ bsfc/bsfc_run_mpi_new.py
@@ -19,16 +19,12 @@ plt.ion()
 import cPickle as pkl
 import pdb
 import corner
-import bsfc_main
 import sys
 import time as time_
 import multiprocessing
 import os
 import shutil
-
-# make it possible to use other packages within the BSFC distribution:
-from os import path
-sys.path.append( path.dirname( path.dirname( path.abspath(__file__) ) ))
+import scipy
 from helpers import bsfc_cmod_shots
 from helpers import bsfc_autocorr
 
@@ -51,7 +47,7 @@ if 'BSFC_ROOT' not in os.environ:
 resume=True
 
 # indicate whether to output verbose text 
-verbose = False
+verbose = True
 
 if rank == 0:
     print "Analyzing shot ", shot
@@ -66,20 +62,20 @@ primary_impurity, primary_line, tbin,chbin, t_min, t_max,tht = bsfc_cmod_shots.g
 
 # Always create new object by default when running with MPI
 if rank==0:
-    mf = bsfc_main.MomentFitter(primary_impurity, primary_line, shot, tht=tht)
+    mf = MomentFitter(primary_impurity, primary_line, shot, tht=tht)
 else:
     mf = None
+
 
 if size==1:
     # if only 1 core is being used, assume that script is being used for plotting
 
-    with open('./bsfc_fits/moments_%d_%dsteps_tmin%f_tmax%f.pkl'%(shot,nsteps,t_min,t_max),'rb') as f:
+    with open('./bsfc_fits/moments_%d_tmin%f_tmax%f.pkl'%(shot,t_min,t_max),'rb') as f:
             gathered_moments=pkl.load(f)
 
     # clean up Hirex-Sr signals -- BR_THRESH might need to be adjusted to eliminate outliers
     moments_vals, moments_stds, time_sel = clean_moments(mf.time, mf.maxChan, t_min,t_max,
                                                          gathered_moments, BR_THRESH=2.0, BR_STD_THRESH=0.1)
-    
     # BSFC slider visualization
     bsfc_slider.visualize_moments(moments_vals, moments_stds, time_sel, q='br')
     bsfc_slider.visualize_moments(moments_vals, moments_stds, time_sel, q='vel')
@@ -104,18 +100,20 @@ else:
     chains_dir = os.path.dirname(basename)
     
     # delete and re-create directory for MultiNest output
-    if os.path.exists(chains_dir):
-        if len(os.listdir(chains_dir))==0:
-            # if directory exists and is empty, everything's ready
-            pass
+    if rank==0:
+        if os.path.exists(chains_dir):
+            if len(os.listdir(chains_dir))==0:
+                # if directory exists and is empty, everything's ready
+                pass
+            else:
+                # directory from previous run exists. Delete and re-create it
+                shutil.rmtree(chains_dir)
+                os.mkdir(chains_dir)
         else:
-            # directory from previous run exists. Delete and re-create it
-            shutil.rmtree(chains_dir)
+            # if directory does not exist, create it
             os.mkdir(chains_dir)
-    else:
-        # if directory does not exist, create it
-        os.mkdir(chains_dir)
-        
+
+    comm.Barrier()
     # map range of channels and compute each
     map_args_tpm = list(itertools.product(range(tidx_min, tidx_max), range(mf.maxChan)))
     map_args = [list(a) for a in map_args_tpm]
@@ -137,43 +135,81 @@ else:
             if not os.path.exists('checkpoints'):
                 os.makedirs('checkpoints')
 
-            if not os.path.exists('checkpoints/%d_%dsteps_tmin%f_tmax%f'%(shot,nsteps,t_min,t_max)):
-                os.makedirs('checkpoints/%d_%dsteps_tmin%f_tmax%f'%(shot,nsteps,t_min,t_max))
+            if not os.path.exists('checkpoints/%d_tmin%f_tmax%f'%(shot,t_min,t_max)):
+                os.makedirs('checkpoints/%d_tmin%f_tmax%f'%(shot,t_min,t_max))
 
             try:
                 # if fit has already been created, re-load it from checkpoint directory
-                resdir ='./checkpoints/%d_%dsteps_tmin%f_tmax%f/moments_%d_%d_bin%d_%d.pkl'%(shot,                                                                                             nsteps,t_min,t_max,shot,nsteps,binn[0],binn[1])
-                with open(resdir,'rb') as f:
-                          res[j] = pkl.load(f)
-                print "Loaded fit moments from ", resdir
+                resfile ='./checkpoints/%d_tmin%f_tmax%f/moments_%d_bin%d_%d.pkl'%(shot,t_min,t_max,shot,binn[0],binn[1])
+                with open(resfile,'rb') as f:
+                    res[j] = pkl.load(f)
+                print "Loaded fit moments from ", resfile
 
             except:
                 # create new fit      
-                mf.fits[binn[0]][binn[1]] = ff(binn)
+                mf.fitSingleBin(tbin=binn[0], chbin=binn[1],NS=True, n_hermite=3, n_live_points=400,
+                                sampling_efficiency=0.3, verbose=verbose)
+                
                 if mf.fits[binn[0]][binn[1]].good ==True:
-                    chain = mf.fits[binn[0]][binn[1]].samples
-                    moments_vals = np.apply_along_axis(mf.fits[binn[0]][binn[1]].lineModel.modelMeasurements, axis=1, arr=chain)
 
-                    res[j] = [np.mean(moments_vals, axis=0), np.std(moments_vals, axis=0)]
+                    # load MultiNest output
+                    mf.fits[binn[0]][binn[1]].NS_analysis(basename)
+                    samples = mf.fits[binn[0]][binn[1]].samples
+                    sample_weights = mf.fits[binn[0]][binn[1]].sample_weights
+                    
+                    # Get model predictions
+                    # \mu = \frac{\sum_i w_i x_i}{\sum_i w_i}
+                    meanw = scipy.einsum('i...,i...->i...', sample_weights,samples).sum(axis=0)/sample_weights.sum(axis=0)
+                    moms=mf.fits[binn[0]][binn[1]].lineModel.modelMeasurements(meanw)
+    
+                    # weighted variance: s^2 = \frac{\sum_i w_i}{(\sum_i w_i)^2 - \sum_i w_i^2}\sum_i w_i (x_i - \mu)^2
+                    V1 = sample_weights.sum(axis=0)
+                    M = scipy.einsum('i...,i...->i...', sample_weights, (samples - meanw)**2).sum(axis=0)
+                    stdw = V1 / (V1**2 - (sample_weights**2).sum(axis=0))*M
+    
+                    # get standard deviation of moments:
+                    moms_up = mf.fits[binn[0]][binn[1]].lineModel.modelMeasurements(meanw+stdw)
+                    moms_down = mf.fits[binn[0]][binn[1]].lineModel.modelMeasurements(meanw-stdw)
+                    moms_std = (moms_up - moms_down)/2.0
+
+                    res[j] = [moms, moms_std]
                 else:
                     res[j] =[np.asarray([np.nan, np.nan, np.nan]), np.asarray([np.nan, np.nan, np.nan])]
 
-                with open('./checkpoints/%d_%dsteps_tmin%f_tmax%f/moments_%d_%d_bin%d_%d.pkl'%(shot,
-                                                                                               nsteps,t_min,t_max,shot,nsteps, binn[0], binn[1]),'wb') as f:
+                with open(resfile,'wb') as f:
                     pkl.dump(res[j],f)
 
         else:
                 
-            mf.fits[binn[0]][binn[1]] = ff(binn)
+            mf.fitSingleBin(tbin=binn[0], chbin=binn[1],NS=True, n_hermite=3, n_live_points=400,
+                                sampling_efficiency=0.3, verbose=verbose)
+       
             if mf.fits[binn[0]][binn[1]].good ==True:
-                chain = mf.fits[binn[0]][binn[1]].samples
-                moments_vals = np.apply_along_axis(mf.fits[binn[0]][binn[1]].lineModel.modelMeasurements, axis=1, arr=chain)
-           
-                res[j] = [np.mean(moments_vals, axis=0), np.std(moments_vals, axis=0)]
+                # load MultiNest output
+                mf.fits[binn[0]][binn[1]].NS_analysis(basename)
+                samples = mf.fits[binn[0]][binn[1]].samples
+                sample_weights = mf.fits[binn[0]][binn[1]].sample_weights
+                
+                # Get model predictions
+                # \mu = \frac{\sum_i w_i x_i}{\sum_i w_i}
+                meanw = scipy.einsum('i...,i...->i...', sample_weights,samples).sum(axis=0)/sample_weights.sum(axis=0)
+                moms=mf.fits[binn[0]][binn[1]].lineModel.modelMeasurements(meanw)
+                
+                # weighted variance: s^2 = \frac{\sum_i w_i}{(\sum_i w_i)^2 - \sum_i w_i^2}\sum_i w_i (x_i - \mu)^2
+                V1 = sample_weights.sum(axis=0)
+                M = scipy.einsum('i...,i...->i...', sample_weights, (samples - meanw)**2).sum(axis=0)
+                stdw = V1 / (V1**2 - (sample_weights**2).sum(axis=0))*M
+                
+                # get standard deviation of moments:
+                moms_up = mf.fits[binn[0]][binn[1]].lineModel.modelMeasurements(meanw+stdw)
+                moms_down = mf.fits[binn[0]][binn[1]].lineModel.modelMeasurements(meanw-stdw)
+                moms_std = (moms_up - moms_down)/2.0
+                
+                res[j] = [moms, moms_std]
             else:
                 res[j] = [np.asarray([np.nan, np.nan, np.nan]), np.asarray([np.nan, np.nan, np.nan])]
-            
-            #pdb.set_trace()
+
+                
     # ===============================
 
     # collect results on rank=0 process:
@@ -187,12 +223,12 @@ else:
         print "*********** Completed fits *************"
     
         # save fits for future use
-        with open('./bsfc_fits/moments_%d_%dsteps_tmin%f_tmax%f.pkl'%(shot,nsteps,t_min,t_max),'wb') as f:
+        with open('./bsfc_fits/moments_%d_tmin%f_tmax%f.pkl'%(shot,t_min,t_max),'wb') as f:
             pkl.dump(gathered_moments, f, protocol=pkl.HIGHEST_PROTOCOL)
 
         if resume:
             # eliminate checkpoint directory if this was created
-            shutil.rmtree('checkpoints/%d_%dsteps_tmin%f_tmax%f'%(shot,nsteps,t_min,t_max))
+            shutil.rmtree('checkpoints/%d_tmin%f_tmax%f'%(shot,t_min,t_max))
         
         # end time count
         elapsed_time = MPI.Wtime() - t_start
